+++ hdt-java-core/src/main/java/org/rdfhdt/hdt/triples/impl/BitmapTriplesIteratorY.java
@@ -95,11 +95,11 @@ public class BitmapTriplesIteratorY implements IteratorTripleID {
 			posZ = prevZ = adjZ.find(posY);
 			nextZ = adjZ.last(posY); 
 			
-			x = (int) adjY.findListIndex(posY)+1;
-			y = (int) adjY.get(posY);
- 			z = (int) adjZ.get(posZ);
+			x = adjY.findListIndex(posY)+1;
+			y = adjY.get(posY);
+ 			z = adjZ.get(posZ);
 		} else {
-			z = (int) adjZ.get(posZ);
+			z = adjZ.get(posZ);
 		}
 		posZ++;	
 	
@@ -129,12 +129,12 @@ public class BitmapTriplesIteratorY implements IteratorTripleID {
 			posZ = prevZ = adjZ.find(posY);
 			nextZ = adjZ.last(posY); 
 			
-			x = (int) adjY.findListIndex(posY)+1;
-			y = (int) adjY.get(posY);
- 			z = (int) adjZ.get(posZ);
+			x = adjY.findListIndex(posY)+1;
+			y = adjY.get(posY);
+ 			z = adjZ.get(posZ);
 		} else {
 			posZ--;
-			z = (int) adjZ.get(posZ);
+			z = adjZ.get(posZ);
 		}
 		
 		updateOutput();
@@ -154,9 +154,9 @@ public class BitmapTriplesIteratorY implements IteratorTripleID {
 		posZ = prevZ = adjZ.find(posY);
 		nextZ = adjZ.last(posY);
 		
-		x = (int) adjY.findListIndex(posY)+1;
-		y = (int) adjY.get(posY);
-        z = (int) adjZ.get(posZ);
+		x = adjY.findListIndex(posY)+1;
+		y = adjY.get(posY);
+        z = adjZ.get(posZ);
 	}
 
 	/* (non-Javadoc)
